DDCNL-9021 | moved /extra-years into its own controller & tests
