+++ app/views/multiyear/transfer/single_year_select.scala.html
@@ -56,7 +56,7 @@ request: AuthenticatedUserRequest[_])
 
     @h1("pages.multi-year-select.heading", elmId = Some("heading"))
 
-    @form(action = routes.TransferController.extraYearsAction(), args = Symbol("id") -> "eligible-years-form") {
+    @form(action = transfer.routes.ExtraYearsController.extraYearsAction(), args = Symbol("id") -> "eligible-years-form") {
     <fieldset class="govuk-fieldset inline spacingBottom" id="earlierYears">
         <div id="year-@availableYears.head.year">
             <h2 class="govuk-heading-m"><span style="color: grey">@Html(Messages("pages.multiyear.taxyear", languageUtilsImpl().taxDateIntervalShort(availableYears.head.year)))</span>
